fix: ensure round_to_precision returns float when precision > 0
When precision > 0, round_to_precision now always returns a float,
even when the input is an integer. Previously, round(2, 2) returned
int(2) instead of float(2.0), causing inconsistent int/float behavior
in Slider and Number components.

Closes #12484

+++ gradio/components/number.py
@@ -124,7 +124,7 @@ class Number(FormComponent):
         elif precision == 0:
             return int(round(num, precision))
         else:
-            return round(num, precision)
+            return float(round(num, precision))
 
     @staticmethod
     def raise_if_out_of_bounds(
